pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.173.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "modules>=1.0.0",
    "openai>=1.90.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "pydantic>=2.11.7",
    "xlsxwriter>=3.2.5",
]
